Update main.py
V

+++ main.py
@@ -38,7 +38,17 @@ async def button(bot, update):
             reply_markup=Script.ABOUT_BUTTONS,
             disable_web_page_preview=True
         )
-
+    if update.data.lower() == "remove":
+        caption = ""
+        await main.edit_message_caption(
+            message_id=message_id, caption=caption, reply_markup=InlineKeyboardMarkup([Data.add_button])
+        )
+    if update.data.lower() == "add":
+        caption = callback_query.message.reply_to_message.caption
+        if caption:
+            await main.edit_message_caption(
+                message_id=message_id, caption=caption, reply_markup=InlineKeyboardMarkup([Data.remove_button])
+            )
         else:
             await update.message.delete()
